Adapt type of node to allow rate-limiting
Since the queueing discipline is a property of the transport, and the
transport had to be created prior to the node, it was not
straightforward to include the node's state in decisions about rate
limiting and backpressure.

To facilitate that, we're changing the type of `node` to not get a
`Transport m`, but instead a function `(LL.Node packing peerData m ->
m (NT.Transport m))`.  That way, we can access the node's state in the
QDisc.