Simplify API again
Changing the type of node to get a `Node m -> m (Transport m)` instead
of a `Transport m` turned out to require quite a complicated refactoring
downstream.

Instead, I defined a new datatype `RateLimiting`, that keeps all the
mutable state needed for rate limiting (the locks), and can be shared by
the Node and the Transport.