Include rate-limiting strategy in tests
This does not yet test the effectiveness of rate-limiting, but it does
test that the tests that we have now do pass when we use a rate-limiting
strategy.

Also, removed the warning logs when a lock was not found, since this is
expected on every first request from a peer (the lock is created upon
receiving the first message).