Provide a configurable delay on incomming connections
We have an existing configurable delay on all incomming
network-transport level messages. This adds a configurable delay on just
incomming lightweight connections. So if this delay is used and the
general one is set to no delay, then we would only get delays on new
lightweight connections. This corresponds to high level / app level
conversations, which in the current app are short-lived. So this delay
more closely corresponds to high level messages.