Change design of packing type
The decoder is decorated with a moand. Motivation: to support a CBOR
packing type which uses IO to decode.

There is now a PackingType class which has two associated types to
determine the packing and unpacking monadic contexts. A packing type
could, for instance, use IO during decoding to get some speedup.

The packing type itself is just used as a type, and is identified by
proxy.

To use a particular packing type within some monad, the Packing record
must be constructed, which shows how to inject the packing and unpacking
monads into the desired monad.