Bring policies in line with latest spec
This is a crucial PR, because in the policy as it had been implemented
requests for block headers and requests for blocks were given the lowest
priority. This would cause relay nodes that are inundated with
transactions (which were given a higher priority) to never request
blocks, therefore getting behind, finally entering recovery mode, and
then being unable to exit recovery mode because transactions were also
given higher priority than the "request headers" messages.

The original spec did not talk about these kinds of messages, but the
newer specs do and indeed give them high priority. This PR brings the
code in line with this spec, but there are still some open questions:

* Do we want to increase the `MaxInFlight` values for core-to-relay,
  relay-to-core, and edge-to-relay communication? Having a value of
  1 causes deadlock unless there are changes made upstream
  (see https://github.com/input-output-hk/cardano-sl/pull/1323)

* The policy docs do not give rules for header requests and block
  requests for edge nodes, and moreover gives transactions the highest
  priority (made sense, since only message considered).  This opens it
  up to the same problems as described above, however; therefore, I've
  given transactions a `PLow` transaction (P4), with requests for block
  headers and requests for blocks a `PHigh` priority (P2). This brings
  the policy here in line with core and relay nodes.
  For exchanges the policy gave P1 for header announcements and P2 for
  transactions; I've left the P1 as is, but here too made the
  transaction P4 and the requests for headers/blocks P2.

* I'm worried about the low max-ahead value of 1 for requests
  for blocks and requests for headers. What happens when a node enters
  recovery mode, and tries to enqueue a whole series of block requests
  because it realizes it's far behind? It will quickly run out of
  peers to encode the requests to. Maybe this is what the "fallback:
  TODO" is about in the policy document.