Use unlimited max-ahead for block requests
This addresses my concern in the previous commit: nodes may need to
queue a whole bunch of requests for blocks. This commit however also
makes a related change: when we have a choice, we pick a node with the
smallest number of messages ahead. This means we at least get some load
balancing when we do request a ton of blocks.